Update portal.c
Updating to allow deed award to all non-immortal present. Base code - if somebody thinks it necessary to ensure they are in a party, that check can be added. Other areas that require other things done will also need an additional check or flag to  make sure the people getting the quest/deed were present for that event.